Check the queries at compile time and other improvements.
query! macro has the advantage to be checked against DB at compile time.
Added migration files with sqlx cli (see readme) to create the contacts table and test table.
So we have contacts for deployment and contacts_t table for tests to avoid conficts.
contacts_t table is created and destroyed with the sqlx::test attribute macro and the test-fixtures script anyway,
but by creating it with sqlx cli as well we avoid having compile errors with the rust-analyzer due to non existing test table.
Add dotenvy to not having to set the environment variable manually.